requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
SQLAlchemy>=2.0
psycopg2-binary>=2.9
python-dotenv>=1.0
pydantic>=2.6
gunicorn>=21.2
